backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python
numpy
statsmodels
websockets
python-dotenv
pydantic
